[fix] 로그인 redux변경 및 오류 수정
